requirements.txt
django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.2
xlwt==1.3.0